+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+### Added
+- Julian Day conversion utilities
+- UTC datetime ↔ Julian Day conversion
+
+## [2.1.0] - 2026-07-23
+
+### Added
+- Rebuilt astronomy constants module
+- Centralized Swiss Ephemeris constants
+- Zodiac, Nakshatra, House and Time constants
+
 ## v1.3.1
 
 ### Added